+++ dx/formatters/dataresource.py
@@ -13,12 +13,7 @@ from pydantic import BaseSettings, Field
 
 from dx.config import DEFAULT_IPYTHON_DISPLAY_FORMATTER, IN_IPYTHON_ENV
 from dx.formatters.main import _register_display_id
-from dx.formatters.utils import (
-    is_default_index,
-    stringify_columns,
-    stringify_indices,
-    truncate_and_describe,
-)
+from dx.formatters.utils import normalize_index_and_columns, truncate_and_describe
 from dx.settings import settings
 
 
@@ -57,27 +52,14 @@ class DXDataResourceDisplayFormatter(DisplayFormatter):
         return DEFAULT_IPYTHON_DISPLAY_FORMATTER.format(obj, **kwargs)
 
 
-def format_dataresource(df: pd.DataFrame, display_id: str) -> tuple:
+def generate_dataresource_body(df: pd.DataFrame, display_id: Optional[str] = None) -> tuple:
     """
     Transforms the dataframe to a payload dictionary containing the
     table schema and column values as arrays.
     """
-    # temporary workaround for numeric column rendering errors with GRID
-    # https://noteables.slack.com/archives/C03CB8A4Z2L/p1658497348488939
-    display_df = df.copy()
-    display_df = stringify_columns(display_df)
-
-    # temporary workaround for numeric MultiIndices
-    # because of pandas build_table_schema() errors
-    if not is_default_index(display_df.index):
-        display_df.reset_index(inplace=True)
-
-    # build_table_schema() also doesn't like pd.NAs
-    display_df.fillna(np.nan, inplace=True)
-
     payload_body = {
-        "schema": build_table_schema(display_df),
-        "data": display_df.reset_index().to_dict("records"),
+        "schema": build_table_schema(df),
+        "data": df.reset_index().to_dict("records"),
         "datalink": {},
     }
     payload = {dataresource_settings.DATARESOURCE_MEDIA_TYPE: payload_body}
@@ -90,16 +72,18 @@ def format_dataresource(df: pd.DataFrame, display_id: str) -> tuple:
     }
     metadata = {dataresource_settings.DATARESOURCE_MEDIA_TYPE: metadata_body}
 
-    if display_id is not None:
-        payload_body["datalink"]["display_id"] = display_id
-        metadata_body["datalink"]["display_id"] = display_id
+    display_id = display_id or str(uuid.uuid4())
+    payload_body["datalink"]["display_id"] = display_id
+    metadata_body["datalink"]["display_id"] = display_id
 
     return (payload, metadata)
 
 
-def _render_dataresource(df, display_id) -> tuple:
+def format_dataresource(df, display_id) -> tuple:
+    # enable 0-n row counts for frontend
+    df = normalize_index_and_columns(df)
     df, dataframe_info = truncate_and_describe(df)
-    payload, metadata = format_dataresource(df, display_id)
+    payload, metadata = generate_dataresource_body(df, display_id)
     metadata[dataresource_settings.DATARESOURCE_MEDIA_TYPE]["datalink"][
         "dataframe_info"
     ] = dataframe_info