+++ dx/tests/test_dataresource.py
@@ -1,13 +1,13 @@
 import uuid
 
-from dx.formatters.dataresource import format_dataresource, get_dataresource_settings
+from dx.formatters.dataresource import generate_dataresource_body, get_dataresource_settings
 
 dataresource_settings = get_dataresource_settings()
 
 
 def test_media_type(sample_dataframe):
     display_id = str(uuid.uuid4())
-    payload, _ = format_dataresource(sample_dataframe, display_id)
+    payload, _ = generate_dataresource_body(sample_dataframe, display_id)
     assert dataresource_settings.DATARESOURCE_MEDIA_TYPE in payload
 
 
@@ -18,7 +18,7 @@ def test_data_structure(sample_dataframe):
     including one for the dataframe's index.
     """
     display_id = str(uuid.uuid4())
-    payload, _ = format_dataresource(sample_dataframe, display_id)
+    payload, _ = generate_dataresource_body(sample_dataframe, display_id)
     data = payload[dataresource_settings.DATARESOURCE_MEDIA_TYPE]["data"]
     assert isinstance(data, list)
     assert len(data) == 3
@@ -31,7 +31,7 @@ def test_data_list_order(sample_dataframe):
     and not column values.
     """
     display_id = str(uuid.uuid4())
-    payload, _ = format_dataresource(sample_dataframe, display_id)
+    payload, _ = generate_dataresource_body(sample_dataframe, display_id)
     data = payload[dataresource_settings.DATARESOURCE_MEDIA_TYPE]["data"]
     assert data[0] == {"col_1": "a", "col_2": "b", "col_3": "c", "index": 0}
     assert data[1] == {"col_1": "a", "col_2": "b", "col_3": "c", "index": 1}
@@ -44,7 +44,7 @@ def test_fields_match_data_width(sample_dataframe):
     the number of dictionary keys per item in the data list.
     """
     display_id = str(uuid.uuid4())
-    payload, _ = format_dataresource(sample_dataframe, display_id)
+    payload, _ = generate_dataresource_body(sample_dataframe, display_id)
     data = payload[dataresource_settings.DATARESOURCE_MEDIA_TYPE]["data"]
     fields = payload[dataresource_settings.DATARESOURCE_MEDIA_TYPE]["schema"]["fields"]
     assert len(data[0]) == len(fields)