+++ dx/tests/test_dx.py
@@ -1,13 +1,13 @@
 import uuid
 
-from dx.formatters.dx import format_dx, get_dx_settings
+from dx.formatters.dx import generate_dx_body, get_dx_settings
 
 dx_settings = get_dx_settings()
 
 
 def test_media_type(sample_dataframe):
     display_id = str(uuid.uuid4())
-    payload, _ = format_dx(sample_dataframe, display_id)
+    payload, _ = generate_dx_body(sample_dataframe, display_id)
     assert dx_settings.DX_MEDIA_TYPE in payload
 
 
@@ -18,7 +18,7 @@ def test_data_structure(sample_dataframe):
     including one for the dataframe's index.
     """
     display_id = str(uuid.uuid4())
-    payload, _ = format_dx(sample_dataframe, display_id)
+    payload, _ = generate_dx_body(sample_dataframe, display_id)
     data = payload[dx_settings.DX_MEDIA_TYPE]["data"]
     assert isinstance(data, list)
     assert len(data) == 4
@@ -31,7 +31,7 @@ def test_data_list_order(sample_dataframe):
     and not row values.
     """
     display_id = str(uuid.uuid4())
-    payload, _ = format_dx(sample_dataframe, display_id)
+    payload, _ = generate_dx_body(sample_dataframe, display_id)
     data = payload[dx_settings.DX_MEDIA_TYPE]["data"]
     assert data[0] == [0, 1, 2]  # index values
     assert data[1] == list("aaa")  # "col_1" values
@@ -45,7 +45,7 @@ def test_fields_match_data_length(sample_dataframe):
     the number of lists in the data list.
     """
     display_id = str(uuid.uuid4())
-    payload, _ = format_dx(sample_dataframe, display_id)
+    payload, _ = generate_dx_body(sample_dataframe, display_id)
     data = payload[dx_settings.DX_MEDIA_TYPE]["data"]
     fields = payload[dx_settings.DX_MEDIA_TYPE]["schema"]["fields"]
     assert len(data) == len(fields)