+++ dx/types.py
@@ -6,6 +6,9 @@ class DXDisplayMode(enum.Enum):
     simple = "simple"  # classic simpleTable/DEX display
     plain = "plain"  # basic/vanilla python/pandas display
 
+    def __str__(self):
+        return str(self.value)
+
 
 class DXSamplingMethod(enum.Enum):
     first = "first"  # df.head(num_rows)
@@ -13,3 +16,6 @@ class DXSamplingMethod(enum.Enum):
     inner = "inner"  # middle rows
     random = "random"  # df.sample(num_rows)
     last = "last"  # df.tail(num_rows)
+
+    def __str__(self):
+        return str(self.value)